[Clang] Move OffloadArch enum to a generic location and add initial set of Intel OffloadArch values
